Disable process crash reports by default
